HW4_VisualQuestionAnswering/requirements.txt
torchtext==0.6.0
tensorboardX
pyyaml
